+++ Sources/Parsing/CollectionStyle.swift
@@ -1,0 +1,10 @@
+extension PureYAML.Parsing {
+    enum CollectionStyle: String, Equatable, CustomStringConvertible {
+        case block
+        case flow
+
+        var description: String {
+            rawValue
+        }
+    }
+}